Moves NewTestingExecutor in testutils package
