+++ internal/terraform/exec/exec_test.go
@@ -3,13 +3,16 @@
 // Copyright (c) 2024 HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package exec
+package exec_test
 
 import (
 	"context"
 	"errors"
 	"testing"
 	"time"
+
+	"github.com/opentofu/tofu-ls/internal/terraform/exec"
+	"github.com/opentofu/tofu-ls/internal/testutils"
 )
 
 func TestExec_timeout(t *testing.T) {
@@ -18,11 +21,11 @@ func TestExec_timeout(t *testing.T) {
 	// See https://github.com/hashicorp/terraform-exec/issues/129
 	t.Skip("upstream implementation prone to race conditions")
 
-	e := NewTestingExecutor(t, t.TempDir())
+	e := testutils.NewTestingExecutor(t, t.TempDir())
 	timeout := 1 * time.Millisecond
 	e.SetTimeout(timeout)
 
-	expectedErr := ExecTimeoutError("Version", timeout)
+	expectedErr := exec.ExecTimeoutError("Version", timeout)
 
 	_, _, err := e.Version(t.Context())
 	if err != nil {
@@ -38,12 +41,12 @@ func TestExec_timeout(t *testing.T) {
 }
 
 func TestExec_cancel(t *testing.T) {
-	e := NewTestingExecutor(t, t.TempDir())
+	e := testutils.NewTestingExecutor(t, t.TempDir())
 
 	ctx, cancelFunc := context.WithCancel(t.Context())
 	cancelFunc()
 
-	expectedErr := ExecCanceledError("Version")
+	expectedErr := exec.ExecCanceledError("Version")
 
 	_, _, err := e.Version(ctx)
 	if err != nil {
